+++ app/templates/dev/scripts/modules/module.js
@@ -4,8 +4,7 @@
 */
 <% if (jsOption ==='RequireJS') { %>define([
     // module dependencies
-    'jquery'
-], function ($) {
+], function () {
 
     'use strict';