+++ test/test-scss-nobootstrap.js
@@ -37,25 +37,7 @@ describe('yeogurt generator SCSS no bootstrap', function () {
             'dev/styles/vendor/_bourbon.scss',
             'dev/styles/vendor/_normalize.scss',
             'dev/styles/print.scss',
-            'dev/styles/main.scss',
-            'dev/dashboard',
-            'dev/dashboard/styles',
-            'dev/dashboard/styles/base',
-            'dev/dashboard/styles/base/_mixins.scss',
-            'dev/dashboard/styles/base/_variables.scss',
-            'dev/dashboard/styles/base/_box-sizing.scss',
-            'dev/dashboard/styles/base/_reset.scss',
-            'dev/dashboard/styles/base/_ie8.scss',
-            'dev/dashboard/styles/partials',
-            'dev/dashboard/styles/partials/_footer.scss',
-            'dev/dashboard/styles/partials/_header.scss',
-            'dev/dashboard/styles/partials/_status-key.scss',
-            'dev/dashboard/styles/partials/_dashboard-switcher.scss',
-            'dev/dashboard/styles/partials/_dashboard.scss',
-            'dev/dashboard/styles/vendor',
-            'dev/dashboard/styles/vendor/_font-awesome.scss',
-            'dev/dashboard/styles/vendor/_normalize.scss',
-            'dev/dashboard/styles/main.scss'
+            'dev/styles/main.scss'
         ];
 
         helpers.mockPrompt(this.app, {
@@ -68,7 +50,7 @@ describe('yeogurt generator SCSS no bootstrap', function () {
             useGA: true,
             useFTP: true,
             jshint: true,
-            extras: ['htaccess', 'useFontAwesome', 'useDashboard']
+            extras: ['htaccess', 'useFontAwesome', 'useBorderBox', 'useDashboard']
         });
         this.app.options['skip-install'] = true;
         this.app.run({}, function () {